make code tidy
