+++ src/handlers/mod.rs
@@ -25,6 +25,7 @@ pub mod install;
 pub mod release;
 pub mod self_;
 pub mod show;
+pub mod switch;
 pub mod update;
 pub mod version;
 pub mod which;
